pyproject.toml
[tool.poetry]
name = "olt"
version = "0.1.0"
description = ""
authors = ["Mike Kushner <mike760807@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
